requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-huggingface
sentence-transformers
chromadb
pypdf
python-dotenv
fastapi
uvicorn
sqlalchemy
alembic
bcrypt
python-jose[cryptography]
pydantic
python-multipart
psycopg2-binary
redis
rq
manim
opencv-python-headless
numpy
google-genai